ml_pipeline/requirements.txt
# ─── ML / NLP ───────────────────────────────────────────────
sentence-transformers==2.7.0
faiss-cpu==1.8.0
scikit-learn==1.4.2
numpy==1.26.4

# ─── PyTorch (CPU) ──────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0

# ─── Data / Visualisation ───────────────────────────────────
matplotlib==3.9.0
seaborn==0.13.2
pandas==2.2.2

# ─── Utilities ──────────────────────────────────────────────
tqdm==4.66.4
joblib==1.4.2
python-dotenv==1.0.1

# ─── Jupyter ────────────────────────────────────────────────
jupyter==1.0.0
notebook==7.2.0
ipykernel==6.29.4